Warning, finishing early, js react files

+++ config/index.js
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 const inquirer = require('inquirer');
 const fs = require('fs-extra');
 
@@ -9,21 +10,32 @@ inquirer
       name: 'environment',
       choices: ['react-typescript', 'react-javascript'],
     },
+    {
+      type: 'confirm',
+      message: 'Files inside public and src folders will be delted. Continue?',
+      name: 'confirmation',
+    },
   ])
-  .then(({ environment, public }) => {
+
+  .then(({ environment, confirmation }) => {
+    if (!confirmation) return;
+
     const cwd = process.cwd();
 
     fs.removeSync(cwd + '/src');
-    console.log('s1');
 
+    // TS
     if (environment === 'react-typescript') {
-      fs.copySync(cwd + '/config/src', cwd + '/src');
+      fs.copySync(cwd + '/config/src/ts', cwd + '/src');
     }
 
+    // TS
+    if (environment === 'react-javascript') {
+      fs.copySync(cwd + '/config/src/js', cwd + '/src');
+    }
+
+    // Public folder
     fs.removeSync(cwd + '/public');
-    console.log('s3');
 
     fs.copySync(cwd + '/config/public', cwd + '/public');
   });
-
-fs;
